move dim and dims to top-level file
