+++ python/ffsim/_slow/fermion_operator.py
@@ -15,8 +15,8 @@ import numpy as np
 import pyscf.fci
 from scipy.sparse.linalg import LinearOperator
 
+from ffsim.dimensions import dim, dims
 from ffsim.operators.fermion_action import FermionAction
-from ffsim.states import dim, dims
 
 
 @dataclass
