+++ python/ffsim/hamiltonians/diagonal_coulomb_hamiltonian.py
@@ -21,8 +21,8 @@ from scipy.sparse.linalg import LinearOperator
 from ffsim import protocols
 from ffsim.contract.diag_coulomb import diag_coulomb_linop
 from ffsim.contract.num_op_sum import num_op_sum_linop
+from ffsim.dimensions import dim
 from ffsim.operators import FermionOperator, cre_a, cre_b, des_a, des_b
-from ffsim.states import dim
 
 
 @dataclasses.dataclass(frozen=True)
