+++ python/ffsim/states/states.py
@@ -12,7 +12,6 @@
 
 from __future__ import annotations
 
-import math
 from collections.abc import Sequence
 from dataclasses import dataclass
 from typing import Optional, cast
@@ -21,6 +20,7 @@ import numpy as np
 import scipy.linalg
 from pyscf.fci.spin_op import contract_ss
 
+from ffsim.dimensions import dim
 from ffsim.states.bitstring import (
     BitstringType,
     addresses_to_strings,
@@ -57,41 +57,6 @@ class StateVector:
         if dtype is None:
             return self.vec
         return self.vec.astype(dtype, copy=False)
-
-
-def dims(norb: int, nelec: tuple[int, int]) -> tuple[int, int]:
-    """Get the dimensions of the FCI space.
-
-    Args:
-        norb: The number of spatial orbitals.
-        nelec: The number of alpha and beta electrons.
-
-    Returns:
-        A pair of integers (dim_a, dim_b) representing the dimensions of the
-        alpha- and beta- FCI space.
-    """
-    n_alpha, n_beta = nelec
-    dim_a = math.comb(norb, n_alpha)
-    dim_b = math.comb(norb, n_beta)
-    return dim_a, dim_b
-
-
-def dim(norb: int, nelec: int | tuple[int, int]) -> int:
-    """Get the dimension of the FCI space.
-
-    Args:
-        norb: The number of spatial orbitals.
-        nelec: Either a single integer representing the number of fermions for a
-            spinless system, or a pair of integers storing the numbers of spin alpha
-            and spin beta fermions.
-
-    Returns:
-        The dimension of the FCI space.
-    """
-    if isinstance(nelec, int):
-        return math.comb(norb, nelec)
-    n_alpha, n_beta = nelec
-    return math.comb(norb, n_alpha) * math.comb(norb, n_beta)
 
 
 # source: pyscf.fci.spin_op.spin_square0
